Cleanup unneeded ResourceType/LogType fields in policy/rule tests
The ResourceType (in policy tests) and LogType (in rule tests) fields are not
necessary since they can be inferred from the policy/rule under test.

The tool has been changed to successfully parse older policy/rule tests that
include these fields.
However, this is a backwards-incompatible change for the analysis-api and
the fields must be present in the tests when deploying them to older Panther versions.

See https://github.com/panther-labs/panther/issues/610 for more context.